pyproject.toml
[tool.poetry]
name = "busy_tally_pipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.1"
pyautogui = "^0.9.54"
opencv-python = "^4.9.0.80"
pylint = "^3.1.0"
psutil = "^5.9.8"
pytest = "^8.1.1"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"
schedule = "^1.2.1"
yagmail = "^0.15.293"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.29"
openpyxl = "^3.1.2"
xlwings = "^0.31.1"
logtail-python = "^0.2.10"
cryptography = "^42.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
